+++ src/components/Sidebar.vue
@@ -48,7 +48,26 @@
 					<HistoryIcon />
 				</template>
 			</NcAppNavigationItem>
-			<NcAppNavigationItem :name="t('news', 'Starred')" icon="icon-starred" :to="{ name: ROUTES.STARRED }">
+			<NcAppNavigationItem
+				:name="t('news', 'Starred')"
+				icon="icon-starred"
+				:to="{ name: ROUTES.STARRED }"
+				:allow-collapse="true"
+				:force-menu="true">
+				<NcAppNavigationItem
+					v-for="group in GroupedStars"
+					:key="group.id"
+					:ref="'starred-' + group.id"
+					:name="group.title"
+					:to="{ name: ROUTES.STARRED, params: { feedId: group.id } }">
+					<template #icon>
+						<RssIcon v-if="!group.faviconLink" />
+						<span v-if="group.faviconLink" style="width: 16px; height: 16px; background-size: contain;" :style="{ backgroundImage: 'url(' + group.faviconLink + ')' }" />
+					</template>
+					<template #counter>
+						<NcCounterBubble :count="group.starredCount" />
+					</template>
+				</NcAppNavigationItem>
 				<template #counter>
 					<NcCounterBubble :count="items.starredCount" />
 				</template>
@@ -406,6 +425,10 @@ export default defineComponent({
 			]
 
 			return navItems
+		},
+
+		GroupedStars(): Array<Feed> {
+			return this.$store.getters.feeds.filter((item: Feed) => item.starredCount !== 0)
 		},
 
 		loading: {