+++ src/components/routes/Starred.vue
@@ -2,10 +2,13 @@
 	<ContentTemplate
 		v-if="!loading"
 		:items="starred"
-		:list-name="t('news', 'Starred')"
-		:list-count="items.starredCount"
-		fetch-key="starred"
-		@load-more="fetchMore()" />
+		:fetch-key="fetchKey"
+		@load-more="fetchMore()">
+		<template #header>
+			{{ t('news', 'Starred') }}
+			<NcCounterBubble class="counter-bubble" :count="feedId ? starred.length : items.starredCount" />
+		</template>
+	</ContentTemplate>
 </template>
 
 <script lang="ts">
@@ -13,6 +16,7 @@ import type { FeedItem } from '../../types/FeedItem.ts'
 
 import { defineComponent } from 'vue'
 import { mapState } from 'vuex'
+import NcCounterBubble from '@nextcloud/vue/components/NcCounterBubble'
 import ContentTemplate from '../ContentTemplate.vue'
 import { ACTIONS, MUTATIONS } from '../../store/index.ts'
 import { outOfScopeFilter, sortedFeedItems } from '../../utils/itemFilter.ts'
@@ -21,12 +25,32 @@ export default defineComponent({
 	name: 'RoutesStarred',
 	components: {
 		ContentTemplate,
+		NcCounterBubble,
+	},
+
+	props: {
+		/**
+		 * Unique identifier of the feed whose starred entries this component displays.
+		 *
+		 * @type {number}
+		 * @example 42 or 0
+		 *
+		 * Used to fetch the relevant entries from the backend and to filter the displayed items.
+		 */
+		feedId: {
+			type: Number,
+			required: false,
+			default: 0,
+		},
 	},
 
 	computed: {
 		...mapState(['items']),
 		starred(): FeedItem[] {
 			let items = this.$store.getters.starred
+			if (this.feedId) {
+				items = this.$store.getters.starred.filter((item: FeedItem) => item.feedId === this.id)
+			}
 			/*
 			 * Sorting items is needed because the allItems array can contain
 			 * different orderings due to possible individual feed sorting
@@ -42,6 +66,14 @@ export default defineComponent({
 		oldestFirst() {
 			return this.$store.getters.oldestFirst
 		},
+
+		id(): number {
+			return Number(this.feedId)
+		},
+
+		fetchKey(): string {
+			return (this.feedId === 0) ? 'starred' : 'starred-' + this.feedId
+		},
 	},
 
 	created() {
@@ -55,8 +87,8 @@ export default defineComponent({
 
 	methods: {
 		async fetchMore() {
-			if (!this.$store.state.items.fetchingItems.starred) {
-				this.$store.dispatch(ACTIONS.FETCH_STARRED)
+			if (!this.$store.state.items.fetchingItems[this.fetchKey]) {
+				this.$store.dispatch(ACTIONS.FETCH_STARRED, { feedId: this.id })
 			}
 		},
 	},